+++ vspace/src/main/webapp/WEB-INF/views/staff/modules/slides/slide.html
@@ -77,9 +77,8 @@
                 	url: "[(@{'/staff/module/'+${module.id}+'/slide/'+${slide.id}+'/space/'})]"+ blockId + "?" +[[${_csrf.parameterName}]]+'='+[[${_csrf.token}]],
     		type: 'GET',
     		success: function(result) {
-    			console.log(result +'hh');
     			$("#selectedSpace").text('');
-    			$("#selectedSpace").text(result);
+    			$("#selectedSpace").text(result.name);
     		    	},
     		error: function(data) {
     			var alert = $('<div class="alert alert-danger alert-dismissible fade show" role="alert"><p>We are sorry but something went wrong. Please try to delete again later.</p><button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button></div>');
